+++ kernel/src/dev/pci/bus.h
@@ -17,14 +17,14 @@ struct pci_bus {
     struct array resources;
     struct spinlock lock;
 
-    uint32_t bus_id;
-    uint32_t segment;
+    uint8_t bus_id;
+    uint8_t segment;
 
     struct list entity_list;
 };
 
 struct pci_bus *
-pci_bus_create(struct pci_domain *domain, uint32_t bus_id, uint32_t segment);
+pci_bus_create(struct pci_domain *domain, uint8_t bus_id, uint8_t segment);
 
 bool pci_add_root_bus(struct pci_bus *bus);
 bool pci_remove_root_bus(struct pci_bus *bus);
